+++ src/pages/Settings/WledCard.jsx
@@ -1,18 +1,8 @@
 import { useEffect } from 'react';
 import { makeStyles } from '@material-ui/core/styles';
-import Card from '@material-ui/core/Card';
-import CardHeader from '@material-ui/core/CardHeader';
-import CardContent from '@material-ui/core/CardContent';
-import InputLabel from '@material-ui/core/InputLabel';
-import MenuItem from '@material-ui/core/MenuItem';
-import { FormControl, Checkbox, Icon, TextField } from '@material-ui/core';
-import Select from '@material-ui/core/Select';
-import Accordion from '@material-ui/core/Accordion';
-import AccordionDetails from '@material-ui/core/AccordionDetails';
-import AccordionSummary from '@material-ui/core/AccordionSummary';
-import Typography from '@material-ui/core/Typography';
+import { Card, CardHeader, CardContent, InputLabel, MenuItem, FormControl, Checkbox, Icon, TextField, Select,  Accordion, AccordionDetails, AccordionSummary, Typography } from '@material-ui/core';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
-import Wled from '../../assets/Wled';
+import Wled from '../../components/Icons/Wled';
 import useStore from '../../utils/apiStore';
 
 const useStyles = makeStyles({
